:bug: fix upload image bug

+++ src/features/riders/controllers/riders.controller.ts
@@ -37,20 +37,16 @@ export class RidersController {
   @Put("me/update-info")
   async updateMe(@User() user: RiderMe, @Body() updateRider: updateRiderDto) {
     if (updateRider.images && updateRider.images.length > 0) {
-      const imagesToKeep = updateRider.images.filter(
-        (image) => !image.isToDelete,
-      );
       const imagesToDelete = updateRider.images.filter(
         (image) => image.isToDelete,
       );
 
-      updateRider.images = await this._ridersService.removeImages(
-        user._id,
-        imagesToKeep,
-      );
+      if (imagesToDelete.length > 0) {
+        await this._ridersService.removeImages(user._id, imagesToDelete);
 
-      for (const image of imagesToDelete) {
-        await this.storageService.deleteImageFromGCP(image.url);
+        for (const image of imagesToDelete) {
+          await this.storageService.deleteImageFromGCP(image.url);
+        }
       }
     }
 
@@ -71,7 +67,6 @@ export class RidersController {
         () => req.files(),
         user.identifier.slug,
       );
-
       if (imagesUrl || imagesUrl.length > 0) {
         await this._ridersService.uploadImages(
           user._id,
